moved deepMerge to util file
